Make sync rollback follow-up fetch converge instead of silently capping
The previous bounded loop in fetchAndMergeSnapshots stopped after 5 passes
with only a warning, leaving optimistic state intact for chained rejections
deeper than the cap. Since fetchedKeys already prevents repeated fetches
and the entity universe is finite, replace the silent early-exit with a
true convergence loop bounded only by a runaway-detection cap that throws.
Extract the shared logic into follow_up_fetch.ts so the HTTP runner and
WebSocket client cannot drift.
